SampDB.go: Initial revision.
First commit for the SampDB module. This commit implements
the basic functionality for the listener, without any actual
database link (local non-volotile storage only).

Inplemented endpoints include:

1. Getting an item from the list (by MAC, IP or Name).
2. Adding item to the list.
3. Assigning an item to an employee (item chosen by
   MAC, IP or Name).
4. Getting the list of items assigned to a specific
   employee.
5. Getting full item list.
6. Getting list of unassigned items.
7. Deleting an item from the list (item chosen by MAC,
IP or Name).

Implementation of admin notification when an item is
assigned to an employee that already has at least two
items assigned is also to be added in further commits.