Add DummyListener service.
In order to be able to test the notifications capabilities
to be introduced to SampDB, add a small utility to mimic the
notifications server and serve as endpoint to notifications
REST accesses.

Note: In order for the two pieces of software to be completely
independent, SampDB is moved in this commit to a subfolder.