SampDB.go, SampDB_test.go, SampDB_io.db: Introduce non-volatile storage and tests.
Add two non-volatile storage mechanisms in addition to the volatile
option. Now when invoking SampDB, the user can specify which storage
option they wish from the options 'volatile', 'json' and 'sqllite'.

The 'volatile' option is the same as before, storing the data in
an in-memory structure.

The 'json' option saves the data in a JSON-formatted text file. The
user can choose which file to use.

The 'sqllite' option saves the data in an SQLLite3 database. Again
the user can choose the filename for the database.

Since now data operations can take longer than before, additionally
a mutex mechanism is introduced to prevent data corruption when
fast requests are generated from multiple sources.

This commit also introduces a test for JSON and a test for SQLite,
as well as repeating all existing test threw all storage mecahnisms.